Deep_learning/requirements.txt
streamlit
tensorflow
numpy
pillow
groq
python-dotenv
matplotlib
requests
